+++ src/main/java/fi/dy/masa/tweakeroo/Tweakeroo.java
@@ -1,7 +1,9 @@
 package fi.dy.masa.tweakeroo;
 
 import fi.dy.masa.malilib.compat.forge.ForgePlatformUtils;
+import fi.dy.masa.tweakeroo.compat.forge.ForgeEventHandler;
 import fi.dy.masa.tweakeroo.gui.GuiConfigs;
+import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
@@ -32,5 +34,7 @@ public class Tweakeroo {
             gui.setParent(screen);
             return gui;
         });
+
+        MinecraftForge.EVENT_BUS.register(new ForgeEventHandler());
     }
 }