+++ src/main/java/fi/dy/masa/tweakeroo/mixin/MixinClientPlayerInteractionManager.java
@@ -45,6 +45,7 @@ public abstract class MixinClientPlayerInteractionManager
         }
     }
 
+    /* Broken with Forge, use `PlayerDestroyItemEvent` to impl
     @Inject(method = "method_41929",
             slice = @Slice(from = @At(value = "INVOKE",
                     target = "Lnet/minecraft/item/ItemStack;use(Lnet/minecraft/world/World;Lnet/minecraft/entity/player/PlayerEntity;Lnet/minecraft/util/Hand;)Lnet/minecraft/util/TypedActionResult;")),
@@ -55,6 +56,7 @@ public abstract class MixinClientPlayerInteractionManager
     {
         PlacementTweaks.onProcessRightClickPost(playerEntity, hand);
     }
+     */
 
     @Inject(method = "interactEntity(" +
             "Lnet/minecraft/entity/player/PlayerEntity;" +