+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ## Change
 - update mafglib to `0.1.9`
+- move `onProcessRightClickPost` from mixin to forge event